Add billing info on useBankCard
