+++ www/src/components/account/billing/BillingBankCards.tsx
@@ -10,7 +10,7 @@ function BillingBankCards() {
   const { card } = useContext(BillingBankCardContext)
   const [edit, setEdit] = useState(false)
 
-  const { error: cardError, renderDisplay, renderEdit } = useBankCard(setEdit, null)
+  const { error: cardError, renderDisplay, renderEdit } = useBankCard(setEdit)
 
   if (cardError) {
     return (
