+++ www/src/components/account/billing/BillingUpgradeToProfessionalModal.tsx
@@ -3,22 +3,15 @@ import {
   useCallback,
   useContext,
   useEffect,
-  useMemo,
   useState,
 } from 'react'
 import { Link } from 'react-router-dom'
 import { useMutation } from '@apollo/client'
 import { Div, Flex } from 'honorable'
-import {
-  Button,
-  FormField,
-  Input,
-  Modal,
-} from '@pluralsh/design-system'
+import { Button, Modal } from '@pluralsh/design-system'
 
 import PlatformPlansContext from '../../../contexts/PlatformPlansContext'
 import BillingBankCardContext from '../../../contexts/BillingBankCardContext'
-import SubscriptionContext from '../../../contexts/SubscriptionContext'
 
 import { SUBSCRIPTION_QUERY, UPGRADE_TO_PROFESSIONAL_PLAN_MUTATION } from './queries'
 
@@ -35,20 +28,8 @@ type BillingUpgradeToProfessionalModalPropsType = {
 function BillingUpgradeToProfessionalModal({ open, onClose }: BillingUpgradeToProfessionalModalPropsType) {
   const { proPlatformPlan, proYearlyPlatformPlan } = useContext(PlatformPlansContext)
   const { card } = useContext(BillingBankCardContext)
-  const { billingAddress } = useContext(SubscriptionContext)
 
   const [applyYearlyDiscount, setApplyYearlyDiscount] = useState(false)
-  const [name, setName] = useState(billingAddress?.name || '')
-  const [line1, setLine1] = useState(billingAddress?.line1 || '')
-  const [line2, setLine2] = useState(billingAddress?.line2 || '')
-  const [city, setCity] = useState(billingAddress?.city || '')
-  const [state, setState] = useState(billingAddress?.state || '')
-  const [zip, setZip] = useState(billingAddress?.zip || '')
-  const [country, setCountry] = useState(billingAddress?.country || '')
-
-  const updatedAddress = useMemo(() => ({
-    name, line1, line2, city, state, zip, country,
-  }), [name, line1, line2, city, state, zip, country])
 
   const [upgradeMutation, { loading }] = useMutation(UPGRADE_TO_PROFESSIONAL_PLAN_MUTATION, {
     variables: {
@@ -63,7 +44,7 @@ function BillingUpgradeToProfessionalModal({ open, onClose }: BillingUpgradeToPr
   const [success, setSuccess] = useState(false)
   const [error, setError] = useState(false)
 
-  const { error: cardError, renderEdit, renderDisplay } = useBankCard(setEdit, updatedAddress, true)
+  const { error: cardError, renderEdit, renderDisplay } = useBankCard(setEdit, true)
 
   const handleUpgrade = useCallback((event: FormEvent) => {
     event.preventDefault()
@@ -72,100 +53,6 @@ function BillingUpgradeToProfessionalModal({ open, onClose }: BillingUpgradeToPr
 
     upgradeMutation()
   }, [card, upgradeMutation])
-
-  const renderBillingForm = useCallback(() => (
-    <>
-      <FormField
-        required
-        label="Full name"
-        marginTop="xsmall"
-      >
-        <Input
-          value={name}
-          onChange={event => setName(event.target.value)}
-          placeholder="Enter full name or company name"
-        />
-      </FormField>
-      <FormField
-        required
-        label="Address line 1"
-        marginTop="xsmall"
-      >
-        <Input
-          value={line1}
-          onChange={event => setLine1(event.target.value)}
-          placeholder="Enter street address"
-        />
-      </FormField>
-      <FormField
-        label="Address line 2"
-        marginTop="xsmall"
-      >
-        <Input
-          value={line2}
-          onChange={event => setLine2(event.target.value)}
-          placeholder="Optional"
-        />
-      </FormField>
-      <FormField
-        required
-        label="City"
-        marginTop="xsmall"
-      >
-        <Input
-          value={city}
-          onChange={event => setCity(event.target.value)}
-          placeholder="Enter city name"
-        />
-      </FormField>
-      <FormField
-        required
-        label="State/Province/Region"
-        marginTop="xsmall"
-      >
-        <Input
-          value={state}
-          onChange={event => setState(event.target.value)}
-          placeholder="Enter state, province, or region"
-        />
-      </FormField>
-      <Flex
-        gap="medium"
-        marginTop="xxsmall"
-      >
-        <FormField
-          required
-          label="Zip/Postal code"
-          flexGrow={1}
-        >
-          <Input
-            value={zip}
-            onChange={event => setZip(event.target.value)}
-            placeholder="Enter zip, or postal code"
-          />
-        </FormField>
-        <FormField
-          required
-          label="Country"
-          flexGrow={1}
-        >
-          <Input
-            value={country}
-            onChange={event => setCountry(event.target.value)}
-            placeholder="Enter country"
-          />
-        </FormField>
-      </Flex>
-    </>
-  ), [
-    name,
-    line1,
-    line2,
-    city,
-    state,
-    zip,
-    country,
-  ])
 
   const renderContent = useCallback(() => (
     <>
@@ -182,7 +69,6 @@ function BillingUpgradeToProfessionalModal({ open, onClose }: BillingUpgradeToPr
       >
         Billing information
       </Div>
-      {!card ? renderBillingForm() : null}
       {edit || !card ? renderEdit() : renderDisplay()}
       <form onSubmit={handleUpgrade}>
         <Flex
@@ -206,7 +92,6 @@ function BillingUpgradeToProfessionalModal({ open, onClose }: BillingUpgradeToPr
     loading,
     renderDisplay,
     renderEdit,
-    renderBillingForm,
     handleUpgrade,
   ])
